SolarEdge: Move coordinators out of sensor platform
